apps/trainer/requirements.txt
pika
python-dotenv
requests
ultralytics